requirements-cuda.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
structlog>=24.1.0
prometheus-client>=0.20.0
httpx>=0.27.0
numpy>=1.26.0
scipy>=1.11.0
# Install torch with CUDA build matching your system; e.g.:
torch==2.2.0+cu118 --index-url https://download.pytorch.org/whl/cu118
botorch>=0.10.0
gpytorch>=1.11.0
scikit-learn>=1.3.0
nats-py>=2.5.0

